`Team Allocation`: Add tutors by searching students
